sprawdzanie czy zasób istnieje

+++ Scripts/Remove-AzureResources.ps1
@@ -1,15 +1,23 @@
-function Remove-resourceGroup {
+function Remove-ResourceGroup {
     param (
         [Parameter(Mandatory = $true)]
         [string]
         $resourceGroupName
     )
 
-    Write-Host "INFO -- Usuwanie grupy zasobów '$resourceGroupName'..."
-    Remove-AzResourceGroup -Name $resourceGroupName -Force -AsJob
+    # Sprawdzenie, czy grupa zasobów istnieje
+    $resourceGroup = Get-AzResourceGroup -Name $resourceGroupName -ErrorAction SilentlyContinue
+
+    if ($null -eq $resourceGroup) {
+        Write-Host "WARN -- Grupa zasobów '$resourceGroupName' nie istnieje."
+    }
+    else {
+        Write-Host "INFO -- Usuwanie grupy zasobów '$resourceGroupName'..."
+        Remove-AzResourceGroup -Name $resourceGroupName -Force -AsJob
+    }
 }
 
-function Remove-serverName {
+function Remove-ServerName {
     param (
         [Parameter(Mandatory = $true)]
         [string]
@@ -20,11 +28,19 @@ function Remove-serverName {
         $serverName
     )
 
-    Write-Host "INFO -- Usuwanie SQL Server '$serverName'..."
-    Remove-AzSqlServer -ResourceGroupName $resourceGroupName -ServerName $serverName -Force
+    # Sprawdzenie, czy serwer SQL istnieje
+    $sqlServer = Get-AzSqlServer -ResourceGroupName $resourceGroupName -ServerName $serverName -ErrorAction SilentlyContinue
+
+    if ($null -eq $sqlServer) {
+        Write-Host "WARN -- Serwer SQL '$serverName' nie istnieje w grupie zasobów '$resourceGroupName'."
+    }
+    else {
+        Write-Host "INFO -- Usuwanie SQL Server '$serverName'..."
+        Remove-AzSqlServer -ResourceGroupName $resourceGroupName -ServerName $serverName -Force
+    }
 }
 
-function Remove-sqlDBName {
+function Remove-SqlDBName {
     param (
         [Parameter(Mandatory = $true)]
         [string]
@@ -39,6 +55,14 @@ function Remove-sqlDBName {
         $serverName
     )
 
-    Write-Host "INFO -- Usuwanie SQL Database '$sqlDBName'..."
-    Remove-AzSqlDatabase -ResourceGroupName $resourceGroupName -ServerName $serverName -DatabaseName $sqlDBName -Force
+    # Sprawdzenie, czy baza danych istnieje
+    $sqlDatabase = Get-AzSqlDatabase -ResourceGroupName $resourceGroupName -ServerName $serverName -DatabaseName $sqlDBName -ErrorAction SilentlyContinue
+
+    if ($null -eq $sqlDatabase) {
+        Write-Host "WARN -- Baza danych '$sqlDBName' nie istnieje na serwerze '$serverName' w grupie zasobów '$resourceGroupName'."
+    }
+    else {
+        Write-Host "INFO -- Usuwanie SQL Database '$sqlDBName'..."
+        Remove-AzSqlDatabase -ResourceGroupName $resourceGroupName -ServerName $serverName -DatabaseName $sqlDBName -Force
+    }
 }
